ref(ui): Refactor dropdown menus into `<DropdownBubble>`
This is the presentational menu that appears when you click something.
