+++ src/sentry/static/sentry/app/components/dropdownAutoCompleteMenu.jsx
@@ -3,13 +3,14 @@ import {Flex} from 'grid-emotion';
 import PropTypes from 'prop-types';
 import React from 'react';
 import _ from 'lodash';
-import styled, {css} from 'react-emotion';
+import styled from 'react-emotion';
 
 import {t} from 'app/locale';
 import AutoComplete from 'app/components/autoComplete';
+import DropdownBubble from 'app/components/dropdownBubble';
 import Input from 'app/views/settings/components/forms/controls/input';
-import space from 'app/styles/space';
 import LoadingIndicator from 'app/components/loadingIndicator';
+import space from 'app/styles/space';
 
 const ItemObjectPropType = {
   value: PropTypes.any,
@@ -403,7 +404,7 @@ class DropdownAutoCompleteMenu extends React.Component {
               })}
 
               {isOpen && (
-                <StyledMenu
+                <BubbleWithMinWidth
                   className={className}
                   {...getMenuProps({
                     ...menuProps,
@@ -462,7 +463,7 @@ class DropdownAutoCompleteMenu extends React.Component {
                       <LabelWithPadding>{renderedFooter}</LabelWithPadding>
                     )}
                   </div>
-                </StyledMenu>
+                </BubbleWithMinWidth>
               )}
             </AutoCompleteRoot>
           );
@@ -471,71 +472,6 @@ class DropdownAutoCompleteMenu extends React.Component {
     );
   }
 }
-
-/**
- * If `blendCorner` is false, then we apply border-radius to all corners
- *
- * Otherwise apply radius to opposite side of `alignMenu`
- */
-const getMenuBorderRadius = ({blendCorner, alignMenu, theme}) => {
-  const radius = theme.borderRadius;
-  if (!blendCorner) {
-    return css`
-      border-radius: ${radius};
-    `;
-  }
-
-  const hasTopLeftRadius = alignMenu !== 'left';
-  const hasTopRightRadius = !hasTopLeftRadius;
-
-  return css`
-    border-radius: ${hasTopLeftRadius ? radius : 0} ${hasTopRightRadius ? radius : 0}
-      ${radius} ${radius};
-  `;
-};
-
-const getMenuArrow = ({menuWithArrow, alignMenu}) => {
-  if (!menuWithArrow) {
-    return '';
-  }
-  const alignRight = alignMenu === 'right';
-
-  return css`
-    top: 32px;
-
-    &::before {
-      width: 0;
-      height: 0;
-      border-left: 9px solid transparent;
-      border-right: 9px solid transparent;
-      border-bottom: 9px solid rgba(52, 60, 69, 0.35);
-      content: '';
-      display: block;
-      position: absolute;
-      top: -9px;
-      left: 10px;
-      z-index: -2;
-      ${alignRight && 'left: auto;'};
-      ${alignRight && 'right: 10px;'};
-    }
-
-    &:after {
-      width: 0;
-      height: 0;
-      border-left: 8px solid transparent;
-      border-right: 8px solid transparent;
-      border-bottom: 8px solid #fff;
-      content: '';
-      display: block;
-      position: absolute;
-      top: -8px;
-      left: 11px;
-      z-index: -1;
-      ${alignRight && 'left: auto;'};
-      ${alignRight && 'right: 11px;'};
-    }
-  `;
-};
 
 const AutoCompleteRoot = styled(({isOpen, ...props}) => <div {...props} />)`
   position: relative;
@@ -629,25 +565,6 @@ const GroupLabel = styled('div')`
   padding: ${space(0.25)} ${space(1)};
 `;
 
-const StyledMenu = styled('div')`
-  background: #fff;
-  border: 1px solid ${p => p.theme.borderDark};
-  position: absolute;
-  top: calc(100% - 1px);
-  min-width: 250px;
-  z-index: ${p =>
-    p.theme.zIndex.dropdownAutocomplete
-      .menu}; /* This is needed to be able to cover e.g. pagination buttons, but also be below dropdown actor button's zindex */
-  right: 0;
-  box-shadow: ${p => p.theme.dropShadowLight};
-  overflow: hidden;
-
-  ${getMenuBorderRadius};
-  ${({alignMenu}) => (alignMenu === 'left' ? 'left: 0;' : '')};
-
-  ${getMenuArrow};
-`;
-
 const StyledItemList = styled('div')`
   max-height: ${p =>
     typeof p.maxHeight === 'number' ? `${p.maxHeight}px` : p.maxHeight};
@@ -661,6 +578,10 @@ const EmptyMessage = styled('div')`
   text-transform: none;
 `;
 
+const BubbleWithMinWidth = styled(DropdownBubble)`
+  min-width: 250px;
+`;
+
 export default DropdownAutoCompleteMenu;
 
-export {StyledMenu, AutoCompleteRoot};
+export {AutoCompleteRoot};
