+++ src/sentry/static/sentry/app/components/dropdownControl.jsx
@@ -1,10 +1,11 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-
 import styled from 'react-emotion';
+
+import DropdownBubble from 'app/components/dropdownBubble';
 import DropdownButton from 'app/components/dropdownButton';
-import MenuItem from 'app/components/menuItem';
 import DropdownMenu from 'app/components/dropdownMenu';
+import MenuItem from 'app/components/menuItem';
 import space from 'app/styles/space';
 
 /*
@@ -34,7 +35,6 @@ class DropdownControl extends React.Component {
   };
 
   static defaultProps = {
-    menuOffset: '39px',
     alwaysRenderMenu: true,
     menuWidth: '100%',
   };
@@ -66,10 +66,11 @@ class DropdownControl extends React.Component {
                 {this.renderButton(isOpen, getActorProps)}
                 <MenuContainer
                   {...getMenuProps({isStyled: true})}
-                  alignRight={alignRight}
-                  menuWidth={menuWidth}
+                  alignMenu={alignRight ? 'right' : 'left'}
+                  width={menuWidth}
                   menuOffset={menuOffset}
                   isOpen={isOpen}
+                  blendCorner
                 >
                   {children}
                 </MenuContainer>
@@ -95,23 +96,10 @@ const StyledDropdownButton = styled(
   font-weight: normal;
 `;
 
-const MenuContainer = styled('ul')`
+const MenuContainer = styled(DropdownBubble.withComponent('ul'))`
   list-style: none;
-  width: ${p => p.menuWidth};
-
-  position: absolute;
-  top: ${p => p.menuOffset};
-  ${p => p.alignRight && 'right: 0'};
   padding: 0;
   margin: 0;
-  z-index: ${p => p.theme.zIndex.dropdownAutocomplete.menu};
-
-  background: ${p => p.theme.background};
-  border-radius: ${p => p.theme.borderRadiusBottom};
-  box-shadow: ${p => p.theme.dropShadowLight};
-  border: 1px solid ${p => p.theme.borderDark};
-  overflow: hidden;
-
   display: ${p => (p.isOpen ? 'block' : 'none')};
 `;
