+++ tests/js/spec/components/settingsBreadcrumbDropdown.spec.jsx
@@ -30,59 +30,59 @@ describe('Settings Breadcrumb Dropdown', function() {
     wrapper.find('Crumb').simulate('mouseEnter');
     jest.runAllTimers();
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(1);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(1);
 
     wrapper.find('Crumb').simulate('mouseLeave');
     // wonder what happens when this arg is negative o_O
     jest.advanceTimersByTime(CLOSE_DELAY - 10);
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(1);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(1);
     jest.advanceTimersByTime(10);
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(0);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(0);
   });
 
   it('closes immediately after selecting an item', function() {
     wrapper.find('Crumb').simulate('mouseEnter');
     jest.runAllTimers();
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(1);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(1);
 
     wrapper
       .find('AutoCompleteItem')
       .first()
       .simulate('click');
-    expect(wrapper.find('StyledMenu')).toHaveLength(0);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(0);
   });
 
   it('stays open when hovered over crumb and then into dropdown menu', function() {
     wrapper.find('Crumb').simulate('mouseEnter');
     jest.runAllTimers();
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(1);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(1);
 
     wrapper.find('Crumb').simulate('mouseLeave');
-    wrapper.find('StyledMenu').simulate('mouseEnter');
+    wrapper.find('BubbleWithMinWidth').simulate('mouseEnter');
     jest.runAllTimers();
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(1);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(1);
   });
 
   it('closes after entering dropdown and then leaving dropdown', function() {
     wrapper.find('Crumb').simulate('mouseEnter');
     jest.runAllTimers();
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(1);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(1);
 
     wrapper.find('Crumb').simulate('mouseLeave');
-    wrapper.find('StyledMenu').simulate('mouseEnter');
+    wrapper.find('BubbleWithMinWidth').simulate('mouseEnter');
     jest.runAllTimers();
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(1);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(1);
 
-    wrapper.find('StyledMenu').simulate('mouseLeave');
+    wrapper.find('BubbleWithMinWidth').simulate('mouseLeave');
     jest.runAllTimers();
     wrapper.update();
-    expect(wrapper.find('StyledMenu')).toHaveLength(0);
+    expect(wrapper.find('BubbleWithMinWidth')).toHaveLength(0);
   });
 });
